NotionLink/requirements.txt
Pillow
pyperclip
pyautogui
notion-client
watchdog
pywin32
sentry-sdk
customtkinter
pystray
